When history max is set, rm failed releases first
Attempts to fix https://github.com/helm/helm/issues/4349

Now "removeLeastRecent", which removed the least recent releases
to make way for new releases, will sort the release history by
version but giving precedence to failed releases, so in the end
all failed releases are at the beginning of the list.

Thus the oldest failed releases are deleted before the newest
failed releases and in the absence of failed releases we delete
the oldest successful releases. This should guarantee that, once
we go above max-history, all releases in the history will be
successful releases that can be safely rollback to (with the
exception of the newest release, which may be failed or not).

This does not take into account other status such as UNKNOWN and
DELETED.